refactor(http-circe): split BodyCodec given into encoder/decoder givens
Deleted CirceCodec.scala and replaced it with CirceInstances.scala,
providing separate circeBodyEncoder and circeBodyDecoder given instances
that preserve decodeAccumulating / List[DecodingError] semantics from #230.
